refactor client/script + move sending onto the event thread
Sending text should properly be on the event thread, since it needs to block
until after the client exits.

After looking at the abortion that was my first attempt to design the
client timeout locking scheme, it occured to me that futures were designed
for this exact reason and I stopped being an idiot and rewrote it.  Now the
class is much clearer and less complex.  Blame fatigue.