major overhaul of text processing
Implemented line and fragment processing in Profile in order to
properly implement the client interface functions, handleLine and
handleFragment.

This necessitated a different, simpler model for TextArea and Text
classes.  Line handling here was improved.

Commands are now attached to the lines they are sent on, even if
multiple are sent, and they're joined with " | " when echoing the
line to the screen.  Since they're saved as separate text, updates
to echo config will be reflected immediately and throughout the
text.

With this, text processing should be extremely predictable to client
code.  Ever prompt is sent immediately as a fragment allowing for
prompt triggers, while full lines are also always sent.

The locking around TextArea was increased, unfortunately, however
it remains possible to reduce the locking overhead long term.  On
the plus side, it's probably free of race conditions, now, except
around scroll position for the top window pane.