copy text works across all text components
Given that we do not want the command line to lose focus, it
becomes somewhat difficult to intercept the copy instruction and
decide when to apply it to a textpane vs a command line.  In that
respect, whatever was highlighted or selected last, will win.

Highlighting itself was inspired by IntelliJ's practice of changing
the background to a pleasing blue.  It may be necessary to allow
the user to configure the highlighting style in the future or change
the background to purple if it's already blue.

Copy does not currently copy color codes.  Doing so would not be
easy and probably not desirable for most users.