added mongodb client, settings, config, events
There is now a MongoConfig, per client, for a mongo db connection,
default is disabled, of course.

There is a settings tab for editing them via the settings window.

Profile has events to manage creation and destruction of the mongo
client.  The mongo client can take some time to initialize (>10s),
so it happens on another thread.  Interaction of this guy with
the java client is going to take some work.

Currently, the client is not used for anything.