fixed connect not to block and now onConnect fires, cleaned up code
Connector was not properly finishing connections, and it turned out
that it was blocking to open the connection, so OP_CONNECT never
fired in the selector.  This ultimately meant the app never printed
connected to the console and it never called the client's onConnect.

Now that it has been fixed, it was also observed that the connect
can still happen without blocking, if the connection is localhost,
so a quickConnect method was added to the Connector interface to
handle when that happens.