iconloader jar, bugfix
Since iconloader is low volatility, it has been changed to a jar and
added to lib.  Eventually, darcula and annotations jar will also live
there.

Iconloader has a bug on Linux where it NPE's when setting up the
Darcula look and feel.  A workaround to this bug has been added to
Main which is only called when the OS is Linux.