fix major connection bug, support telnet GA
Bizarrely, I fucked up and allowed an unknown telnet command to
disconnect the client.  Current behavior, log the unexpected
command.

The command in question was the telnet GA which mud servers use
to indicate a prompt, very nice!

Also, we weren't calling repaint on receiving just a fragment.
This has been fixed, too.